fixed wrong imports,bugs, and refactored imports
