+++ src/App.jsx
@@ -14,15 +14,15 @@ import { auth, db } from "./lib/firebase";
 import { AiOutlineInsertRowAbove } from "react-icons/ai";
 import AnimatedButton from "./components/AnimatedButton";
 import Favorite from "./components/Favorite";
-import Loader from "./components/Loader/Loader";
-import LoginScreen from "./pages/Login/Login";
+import Loader from "./components/Loader";
+import LoginScreen from "./pages/Login";
 import NotFoundPage from "./components/NotFound";
-import Post from "./components/Post/Post";
+import Post from "./components/Post";
 import PostView from "./pages/PostView";
-import Profile from "./pages/Profile/Profile";
+import Profile from "./pages/Profile";
 import ShareModal from "./components/ShareModal";
-import SideBar from "./components/SideBar/SideBar";
-import SignupScreen from "./pages/Signup/Signup";
+import SideBar from "./components/SideBar";
+import SignupScreen from "./pages/Signup";
 import logo from "./assets/logo.png";
 import { makeStyles } from "@mui/styles";
 import { useSnackbar } from "notistack";
