+++ src/components/Camera/index.css
