+++ src/components/ImgUpload/index.jsx
@@ -1,4 +1,4 @@
-import "./ImgUpload.css";
+import "./index.css";
 
 import { Avatar, LinearProgress, TextField } from "@mui/material";
 import { FaChevronCircleLeft, FaChevronCircleRight } from "react-icons/fa";
@@ -6,10 +6,10 @@ import React, { useRef, useState } from "react";
 import { auth, db, handleMultiUpload } from "../../lib/firebase";
 
 import AnimatedButton from "../AnimatedButton";
-import Camera from "../Camera/Camera";
+import Camera from "../Camera";
 import { LazyLoadImage } from "react-lazy-load-image-component";
 import { Link } from "react-router-dom";
-import Popup from "../../reusableComponents/Popup/Popup";
+import Popup from "../../reusableComponents/Popup";
 import firebase from "firebase/compat/app";
 import { useSnackbar } from "notistack";
