+++ src/components/Post/index.css
@@ -48,7 +48,7 @@
       #30e19d,
       #a1d9c9
     ); */
-  background: url(./images/postbg.avif);
+  background: url(../../assets/postbg.avif);
   background-size: cover;
   /* color: rgb(145, 230, 227); */
   color: black;
