+++ src/components/SideBar/index.jsx
@@ -1,5 +1,5 @@
 // custom css import
-import "./SideBar.css";
+import "./index.css";
 
 import { Link, useNavigate } from "react-router-dom";
 
@@ -9,7 +9,7 @@ import { AiOutlineClose } from "react-icons/ai";
 import { Dialog } from "@mui/material";
 import FavoriteBorderIcon from "@mui/icons-material/FavoriteBorder";
 import HomeIcon from "@mui/icons-material/Home";
-import ImgUpload from "../ImgUpload/ImgUpload";
+import ImgUpload from "../ImgUpload";
 import React from "react";
 import { auth } from "../../lib/firebase";
 import { useState } from "react";
