+++ src/pages/PostView.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
 import { doc, getDoc } from "firebase/firestore";
 
-import Loader from "../components/Loader/Loader";
-import Post from "../components/Post/Post";
+import Loader from "../components/Loader";
+import Post from "../components/Post";
 import { db } from "../lib/firebase";
 import { useParams } from "react-router-dom";
