+++ src/reusableComponents/Popup/index.jsx
@@ -1,4 +1,4 @@
-import "./Popup.css";
+import "./index.css";
 
 import React from "react";
